fix: fixes parsing errors
Fixes several parsing errors and changes the type of episodeId to float

fix #6

+++ src/arguments/FilmsArgs.ts
@@ -1,9 +1,9 @@
 import { IsIn, IsNumber } from 'class-validator';
-import { ArgsType, Field, Int } from 'type-graphql';
+import { ArgsType, Field, Float } from 'type-graphql';
 
 @ArgsType()
 export default class FilmArgs {
-	@Field(() => Int, { description: 'The film to look up, referenced by episode number' })
+	@Field(() => Float, { description: 'The film to look up, referenced by episode number' })
 	@IsNumber()
 	@IsIn([1, 2, 3, 4, 5, 6])
 	public film!: number;
